requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
fastapi==0.115.12
uvicorn[standard]==0.34.0

# ── HTTP client (async FCM calls) ─────────────────────────────────────────────
httpx==0.28.1

# ── Firebase / Google Auth ────────────────────────────────────────────────────
google-auth==2.40.1

# ── Data validation ───────────────────────────────────────────────────────────
pydantic==2.11.3

# ── Environment variable loading ──────────────────────────────────────────────
python-dotenv==1.1.0
